[fx] Make add_symlink_to_bin.sh more robust
If .jiri_root/bin doesn't exist, don't try to install into it.

+++ devshell/lib/add_symlink_to_bin.sh
@@ -3,7 +3,10 @@
 # Use of this source code is governed by a BSD-style license that can be
 # found in the LICENSE file.
 
-source "$(cd "$(dirname "${BASH_SOURCE[0]}")" && pwd)"/vars.sh
+devshell_lib_dir="$(cd "$(dirname "${BASH_SOURCE[0]}")" && pwd)"
+FUCHSIA_DIR="$(dirname $(dirname $(dirname "${devshell_lib_dir}")))"
 
-rm -f "${FUCHSIA_DIR}/.jiri_root/bin/fx"
-ln -s "../../scripts/fx" "${FUCHSIA_DIR}/.jiri_root/bin/fx"
+if [[ -d "${FUCHSIA_DIR}/.jiri_root/bin" ]]; then
+  rm -f "${FUCHSIA_DIR}/.jiri_root/bin/fx"
+  ln -s "../../scripts/fx" "${FUCHSIA_DIR}/.jiri_root/bin/fx"
+fi
